app/Http/Requests/Manager/Instructorディレクトリが肥大化しているための整理(yuki)
