+++ app/Http/Requests/Manager/Instructor/Course/IndexRequest.php
@@ -1,12 +1,18 @@
 <?php
 
-namespace App\Http\Requests\Manager;
+namespace App\Http\Requests\Manager\Instructor\Course;
 
-use App\Rules\InstructorIndexSortByRule;
 use Illuminate\Foundation\Http\FormRequest;
 
-class InstructorIndexRequest extends FormRequest
+class IndexRequest extends FormRequest
 {
+    protected function prepareForValidation()
+    {
+        $this->merge([
+            'instructor_id' => $this->route('instructor_id'),
+        ]);
+    }
+
     /**
      * Determine if the user is authorized to make this request.
      *
@@ -25,10 +31,7 @@ class InstructorIndexRequest extends FormRequest
     public function rules()
     {
         return [
-            'per_page' => ['integer', 'min:1', 'max:100'],
-            'page' => ['integer', 'min:1'],
-            'sort_by' => ['string', new InstructorIndexSortByRule],
-            'order' => ['string', 'in:asc,desc'],
+            'instructor_id' => ['required', 'integer', 'exists:instructors,id,deleted_at,NULL'],
         ];
     }
 }
